coffeePoints with coordinates

+++ data/src/main/java/com/example/data/network/models/TokenDTO.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class TokenDTO(
     @SerializedName("token") val token: String,
-    @SerializedName("tokenLifeTime") val tokenLifeTime: Long,
+    @SerializedName("tokenLifeTime") val tokenLifeTime: Int,
 )
